requirements.txt
asgiref==3.7.2
Django==5.0.1
mod-wsgi==5.0.0
paho-mqtt==2.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
